OBPIH-1891 Improve loading performance of dashboard by lazy loading activity summary
